+++ packages/dartvex_local/lib/dartvex_local.dart
@@ -31,6 +31,7 @@ export 'src/client.dart'
         LocalRemoteConnectionState,
         LocalRemoteQueryError,
         LocalRemoteQueryEvent,
+        LocalRemoteQueryLoading,
         LocalRemoteQuerySuccess,
         LocalRemoteSubscription,
         LocalSubscription,