+++ packages/dartvex_local/lib/src/runtime/convex_remote_client.dart
@@ -109,6 +109,8 @@ class _ConvexRemoteSubscriptionAdapter implements LocalRemoteSubscription {
         switch (event) {
           case QuerySuccess(:final value):
             return LocalRemoteQuerySuccess(value);
+          case QueryLoading(:final hasPendingWrites):
+            return LocalRemoteQueryLoading(hasPendingWrites: hasPendingWrites);
           case QueryError(:final message, :final data, :final logLines):
             return LocalRemoteQueryError(
               ConvexException(message, data: data, logLines: logLines),